Move `SelfieSuspend` into its own file.
